+++ components/BannerContent/index.ts
@@ -1,0 +1,3 @@
+import BannerContent from "./component";
+
+export default BannerContent;
