+++ components/BannerContent/styles.module.scss
@@ -1,0 +1,46 @@
+//IMPORTS
+@import '../../styles/base/variables';
+@import '../../styles/function/mixins';
+
+.bannerWrapper{
+  flex: 1;
+  padding: 0 5rem;
+  @include flex(column, center, flex-start, 2.5rem);
+}
+
+.bannerContent{
+  max-width: 500px;
+  @include flex(column, center, flex-start, 2.5rem);
+  h1{
+    font-family: 'Poppins-Bold', sans-serif;
+    line-break: normal;
+  }
+  h3{
+    font-size: 2.8rem;
+    font-family: 'Roboto-Light', sans-serif;
+    line-break: normal;
+  }
+  .callToActionButtons{
+    @include flex(row, center, center, 1rem);
+  }
+}
+
+@media only screen and (max-width: 768px){
+  .bannerWrapper{
+    max-height: calc(90vh / 2);
+    padding: 0 2rem;
+    order: 2;
+    align-self: center;
+    gap: 1.5rem;
+  }
+}
+
+@media only screen and (max-width: 450px){
+  .bannerLeft{
+    .content{
+      padding: 0rem;
+      max-width: 350px;
+    }
+  }
+}
+
