+++ components/BannerImage/component.tsx
@@ -1,0 +1,20 @@
+//STYLES
+import styles from './styles.module.scss';
+
+//TYPES
+type BannerImageProps = {
+  imageUrl?: string;
+}
+
+const BannerImage = ({ imageUrl }: BannerImageProps) => {
+  return(
+    <div 
+      style={{
+        backgroundImage: `url("${imageUrl}")`,
+      }}
+      className={styles.bannerImage} 
+    />
+  );
+}
+
+export default BannerImage;
