+++ components/BannerImage/index.ts
@@ -1,0 +1,3 @@
+import BannerImage from "./component";
+
+export default BannerImage;
